+++ Runtime/Compressors/CompressorFactory.cs
@@ -1,5 +1,3 @@
-using ActionCode.AsyncIO;
-
 namespace ActionCode.Persistence
 {
     /// <summary>
@@ -11,24 +9,12 @@ namespace ActionCode.Persistence
         /// Creates an instance of <see cref="ICompressor"/> based on the given type.
         /// </summary>
         /// <param name="type">The Compressor type to use.</param>
-        /// <returns>An Compressor implementation</returns>
-        public static ICompressor Create(CompressorType type) =>
-            Create(type, StreamFactory.Create());
-
-        /// <summary>
-        /// <inheritdoc cref="Create(CompressorType)"/>
-        /// </summary>
-        /// <param name="type"><inheritdoc cref="Create(CompressorType)"/></param>
-        /// <param name="stream">The stream instance to use.</param>
-        /// <returns><inheritdoc cref="Create(CompressorType)"/></returns>
-        public static ICompressor Create(CompressorType type, IStream stream)
+        /// <returns>A Compressor implementation.</returns>
+        public static ICompressor Create(CompressorType type) => type switch
         {
-            return type switch
-            {
-                CompressorType.None => null,
-                CompressorType.GZip => new GZipCompressor(stream),
-                _ => null
-            };
-        }
+            CompressorType.None => null,
+            CompressorType.GZip => new GZipCompressor(),
+            _ => null
+        };
     }
 }