+++ Runtime/Compressors/ICompressor.cs
@@ -1,4 +1,4 @@
-using System.Threading.Tasks;
+using UnityEngine;
 
 namespace ActionCode.Persistence
 {
@@ -8,14 +8,14 @@ namespace ActionCode.Persistence
         /// Compress the given value.
         /// </summary>
         /// <param name="value">A string to compress.</param>
-        /// <returns>A task operation of the compressing process.</returns>
-        Task<string> Compress(string value);
+        /// <returns>An asynchronous operation of the compressing process.</returns>
+        Awaitable<string> Compress(string value);
 
         /// <summary>
         /// Decompress the given value.
         /// </summary>
         /// <param name="value">A string to decompress.</param>
-        /// <returns>A task operation of the decompressing process.</returns>
-        Task<string> Decompress(string value);
+        /// <returns>An asynchronous operation of the decompressing process.</returns>
+        Awaitable<string> Decompress(string value);
     }
 }