+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Changed
+- Update Async IO package to 2.0.0
 - Update Unity minimum version to 6000.1
 
 ### Removed